revert ss and auditUtl and simplify rt audit
